+++ packages/asc-ui/src/components/Pagination/Pagination.test.tsx
@@ -79,6 +79,36 @@ describe('Pagination', () => {
     expect(screen.getByTestId('pageButton-5')).toHaveAttribute('tabIndex', '-1')
   })
 
+  it('should render customized "previous" and "next" button labels', () => {
+    const onPageChangeMock = jest.fn()
+    render(
+      <Pagination
+        page={6}
+        pageSize={10}
+        collectionSize={100}
+        onPageChange={onPageChangeMock}
+        labelPrevious="Vorige pagina"
+        labelNext="Volgende pagina"
+        ariaLabelPrevious="Vorige aria"
+        ariaLabelNext="Volgende aria"
+      />,
+    )
+
+    expect(screen.getByTestId('nextButton').textContent).toBe('Volgende pagina')
+    expect(screen.getByTestId('previousButton').textContent).toBe(
+      'Vorige pagina',
+    )
+
+    expect(screen.getByTestId('previousButton')).toHaveAttribute(
+      'aria-label',
+      'Vorige aria',
+    )
+    expect(screen.getByTestId('nextButton')).toHaveAttribute(
+      'aria-label',
+      'Volgende aria',
+    )
+  })
+
   it('should be working in a controlled state', () => {
     function ControlledComponent() {
       const [page, setPage] = useState(6)