+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Prefix the change with one of these keywords:
 
 ## Unreleased
 
+- Added: custom labels for Pagination component
 - Added: Support for React 18
 
 ## [0.35.2]